backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flower==2.0.1
humanize==4.14.0
idna==3.11
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
